refactor: moved ThemeProvider into Layout component.
