+++ src/components/layout/layout.tsx
@@ -1,0 +1,27 @@
+import { FC, useMemo } from "react";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+import CssBaseline from "@mui/material/CssBaseline";
+
+type LayoutProps = {
+  isDarkMode: boolean;
+  children: React.ReactNode;
+};
+
+export const Layout: FC<LayoutProps> = ({ isDarkMode, children }) => {
+  const theme = useMemo(
+    () =>
+      createTheme({
+        palette: {
+          mode: isDarkMode ? "dark" : "light",
+        },
+      }),
+    [isDarkMode]
+  );
+
+  return (
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+      {children}
+    </ThemeProvider>
+  );
+};
